requirements.txt
flet==0.25.0
google-genai
python-dotenv